requirements.txt
keyboard
pywin32
Pillow